backend/pyproject.toml
# Backend package metadata, dependencies, and tool config for the
# Scenarios App. See DESIGN.md Section 3 (Tech Stack) and Section 12
# (Configuration) for rationale.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scenarios-app-backend"
version = "0.1.0"
description = "FastAPI backend for the Chinese Scenarios App."
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Scenarios App" }]

# Runtime deps -- pinned only loosely; everyone speaks SemVer well
# enough at these majors that floating minors keeps maintenance cheap.
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "google-genai>=0.3",
    "Pillow>=10.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1",
]

[project.optional-dependencies]
# Pulled in via `pip install -e .[dev]`. Test runner + linter only.
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ─── ruff ─────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
# Selectively enable: pycodestyle (E), Pyflakes (F), import sort (I),
# bugbear (B), pyupgrade (UP). Keep noisy rules off so the linter does
# not turn into background noise.
select = ["E", "F", "I", "B", "UP"]
ignore = [
    "E501", # line length handled by formatter
    "B008", # FastAPI Depends() in defaults is intentional
]

[tool.ruff.lint.per-file-ignores]
# Tests can use slightly looser rules.
"tests/**/*.py" = ["B017"]

# ─── pytest ────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    # SQLAlchemy 2.x emits a DeprecationWarning for datetime.utcnow on
    # Python 3.12+; we use it intentionally for now.
    "ignore::DeprecationWarning",
]
